feat: add relation lookup column type

+++ lib/Db/Row2Mapper.php
@@ -58,6 +58,9 @@ class Row2Mapper {
 		$this->db->beginTransaction();
 		try {
 			foreach ($this->columnsHelper->columns as $columnType) {
+				if ($this->isVirtualColumn($columnType)) {
+					continue;
+				}
 				$this->getCellMapperFromType($columnType)->deleteAllForRow($row->getId());
 			}
 			$this->rowSleeveMapper->deleteById($row->getId());
@@ -195,6 +198,8 @@ class Row2Mapper {
 	private function getRows(array $rowIds, array $columnIds): array {
 		$qb = $this->db->getQueryBuilder();
 
+		$columnIds = $this->addRelationColumnIdsForSupplementColumns($columnIds);
+
 		$qbSqlForColumnTypes = null;
 		foreach ($this->columnsHelper->columns as $columnType) {
 			$qbTmp = $this->db->getQueryBuilder();
@@ -210,6 +215,9 @@ class Row2Mapper {
 				$qbTmp->selectAlias($qbTmp->createFunction('NULL'), 'value_type');
 			}
 
+			if ($this->isVirtualColumn($columnType)) {
+				continue;
+			}
 			$qbTmp
 				->from('tables_row_cells_' . $columnType)
 				->where($qb->expr()->in('column_id', $qb->createNamedParameter($columnIds, IQueryBuilder::PARAM_INT_ARRAY, ':columnIds')))
@@ -798,6 +806,10 @@ class Row2Mapper {
 			throw new InternalError(get_class($this) . ' - ' . __FUNCTION__ . ': ' . $e->getMessage());
 		}
 
+		if ($this->isVirtualColumn($column->getType())) {
+			return;
+		}
+
 		// insert new cell
 		$cellMapper = $this->getCellMapper($column);
 
@@ -836,6 +848,10 @@ class Row2Mapper {
 	 * @throws InternalError
 	 */
 	private function updateCell(RowCellSuper $cell, RowCellMapperSuper $mapper, $value, Column $column): void {
+		if ($this->isVirtualColumn($column->getType())) {
+			return;
+		}
+
 		$this->getCellMapper($column)->applyDataToEntity($column, $cell, $value);
 		$this->updateMetaData($cell);
 		$mapper->updateWrapper($cell);
@@ -846,6 +862,9 @@ class Row2Mapper {
 	 */
 	private function insertOrUpdateCell(int $rowId, int $columnId, $value): void {
 		$column = $this->columnMapper->find($columnId);
+		if ($this->isVirtualColumn($column->getType())) {
+			return;
+		}
 		$cellMapper = $this->getCellMapper($column);
 		try {
 			if ($cellMapper->hasMultipleValues()) {
@@ -872,6 +891,9 @@ class Row2Mapper {
 	}
 
 	private function getCellMapperFromType(string $columnType): RowCellMapperSuper {
+		if ($this->isVirtualColumn($columnType)) {
+			throw new InternalError('Virtual columns do not have cell mappers');
+		}
 		$cellMapperClassName = 'OCA\Tables\Db\RowCell' . ucfirst($columnType) . 'Mapper';
 		/** @var RowCellMapperSuper $cellMapper */
 		try {
@@ -893,6 +915,9 @@ class Row2Mapper {
 	 * @throws InternalError
 	 */
 	public function deleteDataForColumn(Column $column): void {
+		if ($this->isVirtualColumn($column->getType())) {
+			return;
+		}
 		try {
 			$this->getCellMapper($column)->deleteAllForColumn($column->getId());
 		} catch (Exception $e) {
@@ -961,6 +986,9 @@ class Row2Mapper {
 			case Column::TYPE_USERGROUP:
 				$defaultValue = $this->getCellMapper($column)->filterValueToQueryParam($column, $column->getUsergroupDefault());
 				break;
+			case Column::TYPE_RELATION_LOOKUP:
+				$defaultValue = null;
+				break;
 		}
 		return $defaultValue;
 	}
@@ -988,4 +1016,29 @@ class Row2Mapper {
 
 		return $sortedRows;
 	}
+
+	public function isVirtualColumn(string $columnType): bool {
+		return $columnType === Column::TYPE_RELATION_LOOKUP;
+	}
+
+	/**
+	 * @param int[] $columnIds
+	 * @return int[]
+	 */
+	public function addRelationColumnIdsForSupplementColumns(array $columnIds): array {
+		$allColumns = $this->columnMapper->findAll($columnIds);
+		foreach ($allColumns as $column) {
+			if ($column->getType() !== Column::TYPE_RELATION_LOOKUP) {
+				continue;
+			}
+
+			$customSettings = $column->getCustomSettingsArray();
+			$relationColumnId = $customSettings['relationColumnId'] ?? null;
+			if ($relationColumnId && !in_array($relationColumnId, $columnIds)) {
+				$columnIds[] = $relationColumnId;
+			}
+		}
+		return $columnIds;
+	}
+
 }
